requirements/minimal.txt
tensorboard==2.7.0
torch-tb-profiler==0.3.1
ninja==1.10.2
numpy==1.22.3
scipy==1.7.3
scikit-learn==1.0.2
scikit-video==1.1.11
pillow==9.0.0
opencv-python-headless==4.5.5.62
requests
bs4
tqdm
rich
click
cloup
psutil
easydict
lmdb
matplotlib
